# Allow semicolon-separated kwargs in function calls