This fix from my fork seems to port over cleanly to here, and resolves #32, so I figured I'd open a PR. 

Before this PR:

``` julia
julia> @benchmark svds(rand(2,2); nsv = 1)
ERROR: error compiling ##7129: unsupported or misplaced expression "parameters" in function ##7129
 in execute at /Users/jarrettrevels/.julia/v0.4/Benchmarks/src/execute.jl:47
 in execute at /Users/jarrettrevels/.julia/v0.4/Benchmarks/src/execute.jl:42
```

After this PR: 

``` julia
julia> @benchmark svds(rand(2,2); nsv = 1)
================ Benchmark Results ========================
     Time per evaluation: 113.67 μs [44.04 μs, 183.29 μs]
Proportion of time in GC: 0.00% [0.00%, 0.00%]
        Memory allocated: 9.63 kb
   Number of allocations: 170 allocations
       Number of samples: 100
   Number of evaluations: 100
 Time spent benchmarking: 1.90 s
```